If you find the below warning message , apply the fix as described here
"FPU is not initialized, but the project is compiling for an FPU. Please initialize the FPU before use."

So, what is this warning about ?

The build basically warning you that ,you have enabled using hardware floating point unit of the processor in the build settings but you have not initialized that using your code. So, it is warning you to initialize it before using any floating point related code otherwise it may result in processor fault exception.